presenter all
