+++ app/src/main/java/com/example/watchlistpractice/activitiy/DiscoverMoviesActivity.kt
@@ -19,9 +19,7 @@ import com.example.watchlistpractice.R
 import com.example.watchlistpractice.data.ApiData
 import com.example.watchlistpractice.data.RoomMovie
 import com.example.watchlistpractice.fragment.MovieDetailFragment
-import com.example.watchlistpractice.support.ListCardAdapter
-import com.example.watchlistpractice.support.RetrofitInterface
-import com.example.watchlistpractice.support.RoomMovieDatabase
+import com.example.watchlistpractice.support.*
 import kotlinx.android.synthetic.main.activity_discover_movies.*
 import kotlinx.android.synthetic.main.activity_main.*
 import kotlinx.coroutines.CoroutineScope
@@ -30,31 +28,11 @@ import kotlinx.coroutines.launch
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
+import retrofit2.http.OPTIONS
 
-class DiscoverMoviesActivity : AppCompatActivity(), ListCardAdapter.OnMovieListener, ListCardAdapter.DeleteHelper, MovieDetailFragment.OnButtonListener {
-    //  Variables for spinner
-    lateinit var dropdown: Spinner
-    lateinit var spinnerAdapter: ArrayAdapter<String>
-             val OPTIONS: List<String> = listOf("Discover", "Action", "Adventure", "Animation", "Comedy")
+class DiscoverMoviesActivity : AppCompatActivity(), MovieDetailFragment.OnButtonListener {
 
-    //Variable for Retrofit
-    private val RETROFIT_INTERFACE by lazy{
-        RetrofitInterface.create()
-    }
-
-    //Variable to store movies
-    var movieList: ArrayList<RoomMovie> = ArrayList()
-
-    //Variables for recycler view
-    lateinit var recyclerView: RecyclerView
-    lateinit var adapter: ListCardAdapter
-    lateinit var layoutManager: LinearLayoutManager
-
-    //Variable for popup
-    lateinit var popup: PopupWindow
-
-    //Variable for database
-    lateinit var database: RoomMovieDatabase
+    val presenter: DiscoverPresenter = DiscoverPresenter(this)
 
     //Function to show menu
     override fun onCreateOptionsMenu(menu: Menu?): Boolean {
@@ -84,6 +62,10 @@ class DiscoverMoviesActivity : AppCompatActivity(), ListCardAdapter.OnMovieListe
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_discover_movies)
 
+        presenter.setStatusColor()
+
+//        presenter.setToolbar()
+
         bottom_nav_dis.setOnNavigationItemSelectedListener {
             when (it.itemId) {
                 R.id.dis_item_search -> {
@@ -99,137 +81,24 @@ class DiscoverMoviesActivity : AppCompatActivity(), ListCardAdapter.OnMovieListe
             }
         }
 
-        recyclerView = findViewById(R.id.recycler_view_discover_movie)
+        presenter.setRecycler()
 
-        database = Room.databaseBuilder(applicationContext, RoomMovieDatabase::class.java, "data.db").build()
+        presenter.setDatabase()
 
-        dropdown = findViewById(R.id.spinner_genre)
-        spinnerAdapter = ArrayAdapter(this, R.layout.item_genre, OPTIONS)
-        dropdown.adapter = spinnerAdapter
+        presenter.setDropdownSpinner()
 
         button_search_genre.setOnClickListener {
-            movieList = ArrayList()
-
-            val sGenreId: Int? = getGenreId(dropdown.selectedItem.toString())
-
-            val sCall: Call<ApiData.Response> = RETROFIT_INTERFACE.discoverMovie(sGenreId)
-            val sRes: Unit = sCall!!.enqueue(object : Callback<ApiData.Response> {
-                override fun onFailure(call: Call<ApiData.Response>, t: Throwable) {
-                    t.printStackTrace()
-                }
-
-                override fun onResponse(call: Call<ApiData.Response>, response: Response<ApiData.Response>) {
-                    for (movies in response.body()!!.results!!){
-                        movieList.add(RoomMovie(    movies.id!!,
-                                                    movies.title!!,
-                                                    movies.vote_average!!,
-                                                    movies.release_date!!,
-                                                    movies.original_language!!,
-                                                    movies.overview!!,
-                                                    movies.posterPath!!))
-                    }
-                    refreshList()
-                }
-            })
+            presenter.getData()
         }
 
         fab_search_dis.setOnClickListener {
-            movieList = ArrayList()
-
-            val sGenreId: Int? = getGenreId(dropdown.selectedItem.toString())
-
-            val sCall: Call<ApiData.Response> = RETROFIT_INTERFACE.discoverMovie(sGenreId)
-            val sRes: Unit = sCall!!.enqueue(object : Callback<ApiData.Response> {
-                override fun onFailure(call: Call<ApiData.Response>, t: Throwable) {
-                    t.printStackTrace()
-                }
-
-                override fun onResponse(call: Call<ApiData.Response>, response: Response<ApiData.Response>) {
-                    for (movies in response.body()!!.results!!){
-                        movieList.add(RoomMovie(    movies.id!!,
-                            movies.title!!,
-                            movies.vote_average!!,
-                            movies.release_date!!,
-                            movies.original_language!!,
-                            movies.overview!!,
-                            movies.posterPath!!))
-                    }
-                    refreshList()
-                }
-            })
+            presenter.getData()
         }
-    }
-
-    //Function when card is clicked
-    override fun onMovieClick(movie: RoomMovie) {
-        MovieDetailFragment(movie, this, 1).apply {
-            show(supportFragmentManager, tag)
-        }
-//        val customView: View = layoutInflater.inflate(R.layout.popup_movie, null)
-//        popup = PopupWindow(customView, ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT)
-//        popup.isOutsideTouchable = true
-//        popup.isFocusable = true
-//        popup.setBackgroundDrawable(ColorDrawable(Color.LTGRAY))
-//        popup.showAtLocation(relative_layout_activity_discover_movies, Gravity.CENTER, 0, 0)
-//
-//        val popupTextViewTitle: TextView = customView.findViewById(R.id.text_view_title) as TextView
-//        val popupTextViewRating: TextView = customView.findViewById(R.id.text_view_rating) as TextView
-//        val popupTextViewReleased: TextView = customView.findViewById(R.id.text_view_released) as TextView
-//        val popupTextViewLanguage: TextView = customView.findViewById(R.id.text_view_language) as TextView
-//        val popupTextViewDescription: TextView = customView.findViewById(R.id.text_view_description) as TextView
-//        val buttonAddMovie: Button = customView.findViewById(R.id.button_add) as Button
-//
-//        popupTextViewTitle.text = movie.title
-//        popupTextViewRating.text = popupTextViewRating.text.toString() + movie.rating!!
-//        popupTextViewReleased.text = popupTextViewReleased.text.toString() + movie.release!!
-//        popupTextViewLanguage.text = popupTextViewLanguage.text.toString() + movie.language!!
-//        popupTextViewDescription.text = popupTextViewDescription.text.toString() + "\n" + movie.description!!
-//
-//        buttonAddMovie.setOnClickListener{
-//            popup.dismiss()
-//
-//            CoroutineScope(Dispatchers.IO).launch {
-//                addData(movie)
-//            }
-//        }
-    }
-
-    //Function to update the recycler view
-    fun refreshList(){
-        adapter = ListCardAdapter(movieList, this, this)
-
-        layoutManager = LinearLayoutManager(this)
-        recyclerView.layoutManager =  layoutManager
-        recyclerView.adapter = adapter
-    }
-
-    //Function to get genre id
-    fun getGenreId(inGenre: String): Int?{
-        if (inGenre == "Discover") return null
-        else if (inGenre == "Action") return 28
-        else if (inGenre == "Adventure") return 12
-        else if (inGenre == "Animation") return 16
-        else if (inGenre == "Comedy") return 35
-        else return null
-    }
-
-    //Function to add data to local database
-    suspend fun addData(inMovie: RoomMovie){
-        var mChecker = true
-
-        for(movie in database.DataDAO().getData()){
-            if (inMovie.roomMovieId == movie.roomMovieId) mChecker = false
-        }
-        if (mChecker == true) database.DataDAO().insert(inMovie)
-    }
-
-    override fun onSwipe(task: RoomMovie) {
-        TODO("Not yet implemented")
     }
 
     override fun onButtonClick(movie: RoomMovie) {
         CoroutineScope(Dispatchers.IO).launch {
-            addData(movie)
+            presenter.addData(movie)
         }
     }
 }
