+++ app/src/main/java/com/example/watchlistpractice/support/DiscoverPresenter.kt
@@ -1,10 +1,12 @@
 package com.example.watchlistpractice.support
 
+import android.widget.ArrayAdapter
+import android.widget.Spinner
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import androidx.room.Room
 import com.example.watchlistpractice.R
-import com.example.watchlistpractice.activitiy.MainActivity
+import com.example.watchlistpractice.activitiy.DiscoverMoviesActivity
 import com.example.watchlistpractice.data.ApiData
 import com.example.watchlistpractice.data.RoomMovie
 import com.example.watchlistpractice.fragment.MovieDetailFragment
@@ -12,11 +14,15 @@ import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
 
-class MainPresenter (val act: MainActivity): ListCardAdapter.OnMovieListener, ListCardAdapter.DeleteHelper{
+class DiscoverPresenter (val act: DiscoverMoviesActivity): ListCardAdapter.OnMovieListener, ListCardAdapter.DeleteHelper {
 
     private val RETROFIT_INTERFACE by lazy{
         RetrofitInterface.create()
     }
+
+    lateinit var dropdown: Spinner
+    lateinit var spinnerAdapter: ArrayAdapter<String>
+    val OPTIONS: List<String> = listOf("Discover", "Action", "Adventure", "Animation", "Comedy")
 
     lateinit var recyclerView: RecyclerView
     lateinit var adapter: ListCardAdapter
@@ -29,28 +35,42 @@ class MainPresenter (val act: MainActivity): ListCardAdapter.OnMovieListener, Li
     }
 
     fun setRecycler(){
-        recyclerView = act.findViewById(R.id.recycler_view_main)
-    }
-
-    fun setToolbar(){
-        act.setSupportActionBar(act.findViewById(R.id.main_toolbar))
+        recyclerView = act.findViewById(R.id.recycler_view_discover_movie)
     }
 
     fun setDatabase(){
         database = Room.databaseBuilder(act.applicationContext, RoomMovieDatabase::class.java, "data.db").build()
     }
 
-    fun getData(inString: String){
-        val movieList: ArrayList<RoomMovie> = ArrayList()
-        val sCall: Call<ApiData.Response> = RETROFIT_INTERFACE.findMovie(inString)
-        val sRes = sCall!!.enqueue(object: Callback<ApiData.Response> {
+    fun setDropdownSpinner(){
+        dropdown = act.findViewById(R.id.spinner_genre)
+        spinnerAdapter = ArrayAdapter(act, R.layout.item_genre, OPTIONS)
+        dropdown.adapter = spinnerAdapter
+    }
+
+    fun getGenreId(inGenre: String): Int?{
+        if (inGenre == "Discover") return null
+        else if (inGenre == "Action") return 28
+        else if (inGenre == "Adventure") return 12
+        else if (inGenre == "Animation") return 16
+        else if (inGenre == "Comedy") return 35
+        else return null
+    }
+
+    fun getData(){
+        var movieList: ArrayList<RoomMovie> = ArrayList()
+
+        val sGenreId: Int? = getGenreId(dropdown.selectedItem.toString())
+
+        val sCall: Call<ApiData.Response> = RETROFIT_INTERFACE.discoverMovie(sGenreId)
+        val sRes: Unit = sCall!!.enqueue(object : Callback<ApiData.Response> {
             override fun onFailure(call: Call<ApiData.Response>, t: Throwable) {
                 t.printStackTrace()
             }
 
             override fun onResponse(call: Call<ApiData.Response>, response: Response<ApiData.Response>) {
                 for (movies in response.body()!!.results!!){
-                    movieList.add(RoomMovie(movies.id!!,
+                    movieList.add(RoomMovie(    movies.id!!,
                         movies.title!!,
                         movies.vote_average!!,
                         movies.release_date!!,
@@ -89,4 +109,5 @@ class MainPresenter (val act: MainActivity): ListCardAdapter.OnMovieListener, Li
     override fun onSwipe(task: RoomMovie) {
         TODO("Not yet implemented")
     }
+
 }
